# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus Tech Solutions Client Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{id}/milestones:
    get:
      operationId: listProjectMilestones
      tags: [projects]
      summary: List milestones for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"

  # ── TASKS (KANBAN) ───────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/generate:
    post:
      operationId: generateTasks
      tags: [tasks]
      summary: Generate AI tasks for a feature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateTasksRequest"
      responses:
        "201":
          description: Created tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task (move Kanban column, edit, assign)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── RESOURCES ───────────────────────────────────────────────────────────────
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resource consumption entries
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of resource entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"

    post:
      operationId: createResource
      tags: [resources]
      summary: Log a resource consumption entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    patch:
      operationId: updateResource
      tags: [resources]
      summary: Update resource entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourcePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"

    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete resource entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, clientId, status, priority, startDate, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        clientId:
          type: string
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        spent:
          type: ["number", "null"]
        progress:
          type: integer
        assignedDevIds:
          type: array
          items:
            type: string
        githubRepoUrl:
          type: ["string", "null"]
        featuresList:
          type: array
          items:
            $ref: "#/components/schemas/ProjectFeatureItem"
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name, clientId, status, priority, startDate]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        clientId:
          type: string
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        assignedDevIds:
          type: array
          items:
            type: string
    ProjectPatch:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        endDate:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        spent:
          type: ["number", "null"]
        progress:
          type: integer
        assignedDevIds:
          type: array
          items:
            type: string
        githubRepoUrl:
          type: ["string", "null"]
        featuresList:
          type: array
          items:
            $ref: "#/components/schemas/ProjectFeatureItem"
    ProjectFeatureItem:
      type: object
      required: [id, title, completed]
      properties:
        id:
          type: string
        title:
          type: string
        completed:
          type: boolean
    Milestone:
      type: object
      required: [id, projectId, title, dueDate, completed]
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: string
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]

    # ── TASK (KANBAN) ─────────────────────────────────────────────────────────
    Task:
      type: object
      required: [id, title, projectId, status, priority, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        projectId:
          type: string
        featureId:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        assigneeName:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: ["string", "null"]
        estimatedHours:
          type: ["number", "null"]
        loggedHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    GenerateTasksRequest:
      type: object
      required: [projectId, featureId, featureTitle, projectDescription]
      properties:
        projectId:
          type: string
        featureId:
          type: string
        featureTitle:
          type: string
        projectDescription:
          type: string
    TaskInput:
      type: object
      required: [title, projectId, status, priority]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        projectId:
          type: string
        featureId:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: ["string", "null"]
        estimatedHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
    TaskPatch:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        featureId:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: ["string", "null"]
        estimatedHours:
          type: ["number", "null"]
        loggedHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string

    # ── RESOURCE ──────────────────────────────────────────────────────────────
    Resource:
      type: object
      required: [id, projectId, type, quantity, unit, costPerUnit, totalCost, date, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        projectName:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        type:
          type: string
          enum: [dev_hours, hosting, tools, licenses, other]
        description:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        costPerUnit:
          type: number
        totalCost:
          type: number
        date:
          type: string
        createdAt:
          type: string
    ResourceInput:
      type: object
      required: [projectId, type, quantity, unit, costPerUnit, date]
      properties:
        projectId:
          type: string
        type:
          type: string
          enum: [dev_hours, hosting, tools, licenses, other]
        description:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        costPerUnit:
          type: number
        date:
          type: string
    ResourcePatch:
      type: object
      properties:
        projectId:
          type: string
        type:
          type: string
          enum: [dev_hours, hosting, tools, licenses, other]
        description:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        costPerUnit:
          type: number
        date:
          type: string
...