# lib/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Tech Solutions Client Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceInputType } from './resourceInputType';

export interface ResourceInput {
  projectId: string;
  type: ResourceInputType;
  /** @nullable */
  description?: string | null;
  quantity: number;
  unit: string;
  costPerUnit: number;
  date: string;
}