# lib/api-zod/src/generated/types/resourcePatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Tech Solutions Client Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourcePatchType } from './resourcePatchType';

export interface ResourcePatch {
  projectId?: string;
  type?: ResourcePatchType;
  /** @nullable */
  description?: string | null;
  quantity?: number;
  unit?: string;
  costPerUnit?: number;
  date?: string;
}